% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_processing_resolution.R
\name{check_processing_resolution}
\alias{check_processing_resolution}
\title{Check processing resolution of species.}
\usage{
check_processing_resolution(in_species)
}
\arguments{
\item{in_species}{Shapefile of the species being checked for.}
}
\value{
A string with the processing resolution
}
\description{
Checks with which processing resolution the species has 30 or more points.
}
\examples{

}
