% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_presabs.R
\name{create_PA}
\alias{create_PA}
\title{Title}
\usage{
create_PA(input_species_range, target_resolution)
}
\arguments{
\item{input_species_range}{Raster file of species range; usually output of fasterize.}

\item{target_resolution}{Output of check_processing_resolution()}
}
\value{
Returns nothing, writes csv with presence and pseudo-absence coordinates.
}
\description{
Title
}
