% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_presabs.R
\name{create_presabs}
\alias{create_presabs}
\title{Title}
\usage{
create_presabs(input_species_range, target_resolution, write_output = "FALSE")
}
\arguments{
\item{input_species_range}{Simple feature of the species range.}

\item{target_resolution}{Output of check_processing_resolution()}

\item{write_output}{If TRUE, stores the presence-absence dataframe to disk. Default is FALSE.}
}
\value{
Returns nothing, writes csv with presence and pseudo-absence coordinates.
}
\description{
Title
}
